revert(authoring): unwire the half-migrated clip store
~520 lines that nothing called. Verified before removing: no consumer of
`clipEntries`, `clipOrder`, `selectedClipId` or `getAllClips` outside the
store's own files, and the only callers of `addClip` and friends were
tests.

Finishing the migration instead would have meant mirroring
`authoredAnimationTargets` plus three override maps into `clipEntries` on
every change — two sources of truth kept in step by an effect, which is
precisely the shape of the corruption this branch spent most of its time
fixing, and which two earlier attempts on this branch already produced
and reverted. The export memo also encodes a rule the store has no notion
of: an imported clip matching its baseline with no overrides is excluded
from export.

Single-owner is still the right destination. It belongs with the App.tsx
decomposition, where App *stops* holding clips so there is nothing to
mirror — not smuggled in behind bug fixes in a 107-file PR.

Kept: `nextClipOrdinal` moves to `utils/animationClipIds.ts`. It is live
in three call sites and guards the clip-id collision that let one clip's
tracks overwrite another's; its docblock and test come with it.

Dropped with the store: `materialiseBuffer`, and the reset/resetAll split
added two commits ago — with no clip set in `INITIAL_STATE` the two are
identical again, and the hazard it defused no longer exists.

The reducers and their 12 tests encode real decisions about clip
identity, ordering and selection. Recoverable at 1ad692dc rather than
re-derived.
